+++ changelog.md
@@ -5,6 +5,11 @@
 - changed `quiz_option_component` method to return the element instead of adding it to the parent directly
 - moved the add option button creation and event binding outside of the `quiz_option_component` method  
 - implemented the add option button to insert new option component before itself when clicked
+- added the last option add button to be created after all existing options are created
+- changed `QuizPlayer` to work with multiple correct options
+- changed reset button in `QuizForm` to confirm selected options instead of resetting the quiz
+- changed score updating to give partial score based on number of correct options selected over total selected options 
+
 
 # 13-11-2025
 - added option add button to without functionality